requirements.txt
allure-python-commons==2.13.2
async-generator==1.10
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
py==1.11.0
PySocks==1.7.1
pytest==7.3.2
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-selenium==4.0.1
pytest-variables==3.0.0
requests==2.31.0
selenium==4.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==8.2.2
termcolor==2.3.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
wsproto==1.2.0

python-dotenv~=1.0.0
pendulum~=2.1.2